+++ packages/eas-cli/src/submit/android/__tests__/AndroidSubmitCommand-test.ts
@@ -20,7 +20,7 @@ import {
 import { SubmissionMutation } from '../../../graphql/mutations/SubmissionMutation';
 import { createTestProject } from '../../../project/__tests__/project-utils';
 import { getOwnerAccountForProjectIdAsync } from '../../../project/projectUtils';
-import { getVcsClient } from '../../../vcs';
+import { resolveVcsClient } from '../../../vcs';
 import {
   ArchiveResolverContext,
   ArchiveSource,
@@ -57,7 +57,7 @@ jest.mock('../../commons', () => {
   };
 });
 
-const vcsClient = getVcsClient();
+const vcsClient = resolveVcsClient();
 
 describe(AndroidSubmitCommand, () => {
   const testProject = createTestProject(testProjectId, mockJester.accounts[0].name, {