+++ packages/eas-cli/src/submit/ios/__tests__/AscApiKeySource-test.ts
@@ -17,7 +17,7 @@ import { AppQuery } from '../../../graphql/queries/AppQuery';
 import { createTestProject } from '../../../project/__tests__/project-utils';
 import { getBundleIdentifierAsync } from '../../../project/ios/bundleIdentifier';
 import { promptAsync } from '../../../prompts';
-import { getVcsClient } from '../../../vcs';
+import { resolveVcsClient } from '../../../vcs';
 import { SubmissionContext, createSubmissionContextAsync } from '../../context';
 import {
   AscApiKeySource,
@@ -42,7 +42,7 @@ const testProject = createTestProject(testProjectId, mockJester.accounts[0].name
 });
 const projectId = uuidv4();
 
-const vcsClient = getVcsClient();
+const vcsClient = resolveVcsClient();
 
 async function getIosSubmissionContextAsync(): Promise<SubmissionContext<Platform.IOS>> {
   const graphqlClient = instance(mock<ExpoGraphqlClient>());